requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# ML/DL
torch>=2.1.0
pytorch-lightning>=2.1.0
transformers>=4.35.0

# Bioinformatics
biopython>=1.81
biotite>=0.37.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
py3Dmol>=2.0.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.1.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
kaggle>=1.5.16
pyyaml>=6.0.1

# Experiment tracking
wandb>=0.16.0
